feat(tui): remove top bar, add infinite scrollback, global tags
- Remove top bar, move model/token/cost info to global tags line
- Global tags shows status when running, model info when idle
- Implement infinite scrollback with VT100 escape sequences
- Completed turns pushed to terminal scrollback, not viewport
- Layout: [feed, global_tags, input, hotkeys]
- Add ContextState replacing TopBarState
- 398 tests passing

Refs: codex-style infinite scrolling

+++ crates/runie-tui/src/tui/tests/reducer/agent_tests.rs
@@ -24,7 +24,7 @@ fn test_agent_event_message_start() {
     assert_eq!(state.messages.len(), 1);
 }
 
-fn create_message_start_event(turn: u32) -> AgentEvent {
+fn create_message_start_event(turn: usize) -> AgentEvent {
     AgentEvent::MessageStart {
         message: AgentMessage {
             role: "assistant".to_string(),
@@ -39,7 +39,7 @@ fn create_message_start_event(turn: u32) -> AgentEvent {
     }
 }
 
-fn create_message_update_event(turn: u32, text: &str) -> AgentEvent {
+fn create_message_update_event(turn: usize, text: &str) -> AgentEvent {
     use runie_agent::ContentPart;
     AgentEvent::MessageUpdate {
         message: AgentMessage {
